issue #177: adapt to base setup
