+++ app/constants.py
@@ -1,15 +1,15 @@
-# import os
+import os
 
-# from dotenv import load_dotenv
+from dotenv import load_dotenv
 
-# load_dotenv()
+load_dotenv()
 
 # API_ENDPOINT = os.getenv("AZURE_API_ENDPOINT")
 # API_KEY = os.getenv("AZURE_API_KEY")
 # BASE_PATH = os.getenv("API_BASE_PATH", "")
 # FRONTEND_URL = os.getenv("FRONTEND_URL", "*")
-# FERTISCAN_DB_URL = os.getenv("FERTISCAN_DB_URL")
-# FERTISCAN_SCHEMA = os.getenv("FERTISCAN_SCHEMA")
+FERTISCAN_DB_URL = os.getenv("FERTISCAN_DB_URL")
+FERTISCAN_SCHEMA = os.getenv("FERTISCAN_SCHEMA")
 # FERTISCAN_STORAGE_URL = os.getenv("FERTISCAN_STORAGE_URL")
 # OPENAI_API_DEPLOYMENT = os.getenv("AZURE_OPENAI_DEPLOYMENT")
 # OPENAI_API_ENDPOINT = os.getenv("AZURE_OPENAI_ENDPOINT")
